requirements.txt
asgiref==3.4.1
attrs==21.4.0
cffi==1.15.0
cryptography==36.0.2
Django==3.2.12
django-debug-toolbar==3.2.4
django-filter==21.1
djangorestframework==3.13.1
drf-spectacular==0.22.0
importlib-metadata==4.8.3
inflection==0.5.1
jsonschema==3.2.0
Pillow==8.4.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
zipp==3.6.0
